[REFACTOR] Factorise live progress report assessments
